# docs/contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: PlayNexus MVP API
  version: 1.0.0
  summary: Tenant-safe operational API for PlayNexus
  description: >-
    Versioned REST contract for the PlayNexus Phase 1 MVP. Tenant context is
    derived from the authenticated identity; tenant_id is never accepted as a
    writable field on tenant-owned resources. Money is expressed in minor
    units and all timestamps are UTC RFC 3339 values.
servers:
  - url: https://api.example.com/api/v1
    description: Environment-specific base URL
...
security:
  - bearerAuth: []
paths:
...
  /pos/orders:
    post:
      tags: [POS]
      operationId: createOrder
      summary: Create a priced POS order
      description: 'Permission: orders.manage. Client prices are not accepted.'
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '201':
          description: Open or pending-payment order
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
...
components:
  securitySchemes:
    webSession:
      type: apiKey
      in: cookie
      name: playnexus_session
      description: First-party Laravel session cookie (configured cookie name may vary); state-changing web requests additionally require CSRF verification.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
      description: Revocable Laravel Sanctum token returned once at login.
    providerSignature:
      type: apiKey
      in: header
      name: X-Provider-Signature
      description: Provider-specific HMAC/signature over the unmodified body and timestamp; verified before payload processing.
  headers:
    ETag:
      description: Quoted aggregate lock version.
      schema:
        type: string
        pattern: '^"[0-9]+"$'
        example: '"3"'
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Unique command key scoped to actor, tenant, and operation.
      schema:
        type: string
        minLength: 16
        maxLength: 100
        pattern: '^[!-~]+$'
...
  responses:
...
    Unauthenticated:
      description: Authentication is missing, invalid, or expired
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://docs.playnexus.example/problems/unauthenticated
            title: Authentication required
            status: 401
            code: unauthenticated
            detail: A valid bearer token is required.
            instance: /api/v1/auth/me
            request_id: b7b7629e-1259-4ee0-a65e-262419c20ef0
    Forbidden:
      description: The actor lacks permission for this action
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource does not exist or is outside the visible scope
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Request conflicts with current state, version, capacity, or idempotency record
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ValidationFailed:
      description: Request is syntactically valid JSON but fails semantic validation
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
    Ulid:
      type: string
      pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
      example: 01K3F26MGG34P4KZVY8W2RQ1AX
    Currency:
      type: string
      pattern: '^[A-Z]{3}$'
      example: EGP
    LockVersion:
      type: integer
      minimum: 1
      example: 1
...
    OrderStatus:
      type: string
      enum: [draft, paid, refunded, voided]
...
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field:
          type: string
          description: JSON Pointer to the invalid input.
        code:
          type: string
        message:
          type: string
    Problem:
      type: object
      required: [type, title, status, code, detail, instance, request_id]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          pattern: '^[a-z][a-z0-9_]*$'
        detail:
          type: string
        instance:
          type: string
        request_id:
          type: string
          format: uuid
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
...
    OrderItem:
      type: object
      required:
        - id
        - line_number
        - item_kind
        - description
        - quantity
        - unit_price_minor
        - discount_minor
        - tax_minor
        - line_total_minor
        - currency
      properties:
        id:
          $ref: '#/components/schemas/Ulid'
        line_number:
          type: integer
          minimum: 1
        item_kind:
          type: string
          enum: [product, ticket, session, manual_add_on]
        product_id:
          $ref: '#/components/schemas/Ulid'
        ticket_type_id:
          $ref: '#/components/schemas/Ulid'
        session_id:
          $ref: '#/components/schemas/Ulid'
        description:
          type: string
        quantity:
          type: integer
          minimum: 1
        unit_price_minor:
          type: integer
        discount_minor:
          type: integer
          minimum: 0
        tax_minor:
          type: integer
        line_total_minor:
          type: integer
        currency:
          $ref: '#/components/schemas/Currency'
    Order:
      type: object
      required:
        - id
        - branch_id
        - status
        - items
        - subtotal_minor
        - discount_minor
        - tax_minor
        - total_minor
        - paid_minor
        - refunded_minor
        - currency
        - lock_version
        - created_at
      properties:
        id:
          $ref: '#/components/schemas/Ulid'
        branch_id:
          $ref: '#/components/schemas/Ulid'
        guardian_id:
          oneOf:
            - $ref: '#/components/schemas/Ulid'
            - type: 'null'
        session_id:
          oneOf:
            - $ref: '#/components/schemas/Ulid'
            - type: 'null'
        receipt_number:
          type: [string, 'null']
        status:
          $ref: '#/components/schemas/OrderStatus'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        subtotal_minor:
          type: integer
        discount_minor:
          type: integer
          minimum: 0
        tax_minor:
          type: integer
        total_minor:
          type: integer
        paid_minor:
          type: integer
          minimum: 0
        refunded_minor:
          type: integer
          minimum: 0
        currency:
          $ref: '#/components/schemas/Currency'
        lock_version:
          $ref: '#/components/schemas/LockVersion'
        created_at:
          type: string
          format: date-time
        completed_at:
          type: [string, 'null']
          format: date-time
    OrderCreateItem:
      type: object
      additionalProperties: false
      required: [item_kind, quantity]
      properties:
        item_kind:
          type: string
          enum: [product, ticket, session, manual_add_on]
        product_id:
          $ref: '#/components/schemas/Ulid'
        ticket_type_id:
          $ref: '#/components/schemas/Ulid'
        session_id:
          $ref: '#/components/schemas/Ulid'
        quantity:
          type: integer
          minimum: 1
          maximum: 1000
        manual_description:
          type: string
          maxLength: 255
        manual_unit_price_minor:
          type: integer
          minimum: 0
    OrderCreateRequest:
      type: object
      additionalProperties: false
      required: [branch_id, items]
      properties:
        branch_id:
          $ref: '#/components/schemas/Ulid'
        guardian_id:
          $ref: '#/components/schemas/Ulid'
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/OrderCreateItem'
        discount_minor:
          type: integer
          minimum: 0
          default: 0
        discount_reason:
          type: string
          maxLength: 500
        approval_id:
          $ref: '#/components/schemas/Ulid'
...